requirements.txt
torch
torchvision
h5py
matplotlib
numpy
pillow
